+++ nxp/project/pwm.h
@@ -15,7 +15,7 @@ static volatile unsigned int PWMTick3 = 0;
 void SetDutyCycle0(unsigned int DutyCycle, unsigned int Frequency, int dir);
 void InitPWM0(void);
 void PWM_ISR0(void);
-void SetDutyCycle3(unsigned int DutyCycle, unsigned int Frequency);
+void SetDutyCycle3(float DutyCycle, unsigned int Frequency);
 void SetDutyCycle0(unsigned int DutyCycle, unsigned int Frequency, int dir);
 void InitPWM3(void);
 void PWM_ISR3(void);
